requirements.txt
flet
psycopg2-binary
python-dotenv